+++ src/parsers.js
@@ -9,8 +9,7 @@ export const parseBasicRoom = data => ({
   id: data.id,
   isPrivate: data.private,
   name: data.name,
-  updatedAt: data.updated_at,
-  userIds: data.member_user_ids
+  updatedAt: data.updated_at
 })
 
 export const parseBasicUser = data => ({
@@ -23,9 +22,7 @@ export const parseBasicUser = data => ({
 })
 
 export const parsePresence = data => ({
-  lastSeenAt: data.last_seen_at,
-  state: contains(data.state, ['online', 'offline']) ? data.state : 'unknown',
-  userId: data.user_id
+  state: contains(data.state, ['online', 'offline']) ? data.state : 'unknown'
 })
 
 export const parseBasicMessage = data => ({