+++ src/room-store.js
@@ -1,14 +1,15 @@
-import { append, map, filter, uniq, curry, pipe } from 'ramda'
+import { append, map, filter, uniq, pipe } from 'ramda'
 
 import { Store } from './store'
 import { parseBasicRoom } from './parsers'
 import { Room } from './room'
 
 export class RoomStore {
-  constructor ({ instance, userStore, logger }) {
-    this.instance = instance
-    this.userStore = userStore
-    this.logger = logger
+  constructor (options) {
+    this.instance = options.instance
+    this.userStore = options.userStore
+    this.isSubscribedTo = options.isSubscribedTo
+    this.logger = options.logger
   }
 
   store = new Store()
@@ -17,46 +18,59 @@ export class RoomStore {
     this.store.initialize(map(this.decorate, initial))
   }
 
-  set = curry((roomId, basicRoom) => {
+  set = (roomId, basicRoom) => {
+    const room = this.store.getSync(roomId)
+    if (room) {
+      return Promise.resolve(room)
+    }
     return this.store.set(roomId, this.decorate(basicRoom))
-      .then(room =>
-        this.userStore.fetchMissingUsers(room.userIds).then(() => room)
-      )
-  })
+  }
 
   get = roomId => this.store.get(roomId).then(room =>
-    room || this.fetchBasicRoom(roomId).then(this.set(roomId))
+    room || this.fetchBasicRoom(roomId).then(basicRoom => this.set(roomId, basicRoom))
   )
 
   pop = this.store.pop
 
-  addUserToRoom = (roomId, userId) => this.store.update(roomId, r => {
-    r.userIds = uniq(append(userId, r.userIds))
-    return r
-  })
+  addUserToRoom = (roomId, userId) => {
+    return Promise.all([
+      this.store.update(roomId, r => {
+        r.userIds = uniq(append(userId, r.userIds))
+        return r
+      }),
+      this.userStore.fetchMissingUsers([userId])
+    ])
+      .then(([room]) => room)
+  }
 
   removeUserFromRoom = (roomId, userId) => this.store.update(roomId, r => {
     r.userIds = filter(id => id !== userId, r.userIds)
     return r
   })
 
-  update = (roomId, updates) => this.store.update(roomId, r => {
-    r.createdAt = updates.createdAt || r.createdAt
-    r.createdByUserId = updates.createdByUserId || r.createdByUserId
-    r.deletedAt = updates.deletedAt || r.deletedAt
-    r.id = updates.id || r.id
-    r.isPrivate = updates.isPrivate || r.isPrivate
-    r.name = updates.name || r.name
-    r.updatedAt = updates.updatedAt || r.updatedAt
-    r.userIds = updates.userIds || r.userIds
-    return r
-  })
+  update = (roomId, updates) => {
+    return Promise.all([
+      this.store.update(roomId, r => {
+        r.createdAt = updates.createdAt || r.createdAt
+        r.createdByUserId = updates.createdByUserId || r.createdByUserId
+        r.deletedAt = updates.deletedAt || r.deletedAt
+        r.id = updates.id || r.id
+        r.isPrivate = updates.isPrivate || r.isPrivate
+        r.name = updates.name || r.name
+        r.updatedAt = updates.updatedAt || r.updatedAt
+        r.userIds = updates.userIds || r.userIds
+        return r
+      }),
+      this.userStore.fetchMissingUsers(updates.userIds || [])
+    ])
+      .then(([room]) => room)
+  }
 
   fetchBasicRoom = roomId => {
     return this.instance
       .request({
         method: 'GET',
-        path: `/rooms/${roomId}`
+        path: `/rooms/${encodeURIComponent(roomId)}`
       })
       .then(pipe(JSON.parse, parseBasicRoom))
       .catch(err => {
@@ -70,7 +84,12 @@ export class RoomStore {
 
   decorate = basicRoom => {
     return basicRoom
-      ? new Room(basicRoom, this.userStore)
+      ? new Room({
+        basicRoom,
+        userStore: this.userStore,
+        isSubscribedTo: this.isSubscribedTo,
+        logger: this.logger
+      })
       : undefined
   }
 }